Changes rating data type to float.

+++ database/migrations/2019_07_26_105254_create_user_track_rating_table.php
@@ -16,7 +16,7 @@ class CreateUserTrackRatingTable extends Migration
         Schema::create('user_track_rating', function (Blueprint $table) {
             $table->unsignedInteger('user_id');
             $table->unsignedInteger('track_id');
-            $table->unsignedSmallInteger('rating');
+            $table->float('rating');
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users');
